backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
qdrant-client==1.12.1
openai==1.52.2
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
asyncpg==0.29.0
better-exceptions==0.3.3
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
tiktoken==0.8.0
sentence-transformers==3.1.1
torch==2.5.0
datasets==3.0.2
accelerate==0.34.2
transformers==4.45.2
textstat==0.7.3